requirements.txt
alembic==1.8.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
haversine==2.6.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
Mako==1.2.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.2
requests==2.28.1
scikit-learn==1.0.2
scipy==1.9.0
six==1.16.0
SQLAlchemy==1.4.39
threadpoolctl==3.1.0
torch==1.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.1.2
WTForms==3.0.1
xgboost==0.90
